+++ ChangeLog.md
@@ -1,5 +1,12 @@
 # Wasat ChangeLog
 
+## Unreleased
+
+**Released: WiP**
+
+- Added `GeminiURI.without_query`.
+  ([#32](https://github.com/davep/wasat/pull/32))
+
 ## v0.7.0
 
 **Released: 2026-07-16**